Add Vertex for Triangles. Implement normal interp.
I changed Triangle so it holds a Vertex containing a position and
normal, instead of having triangles hold only their face normals. This
allowed me to implement normal interpolation for meshes and have smooth
surfaces even with low triangle-count meshes. See the output folder for
a before and after example with the Utah teapot.